+++ extensions/src/index.ts
@@ -1,4 +1,5 @@
 export * from './extensions/getSearch';
+export * from './extensions/getSelectedOptions';
 export * from './extensions/getValue';
 export * from './extensions/setReadonly';
 export * from './extensions/setValue';